Update wpt results for the passing scroll test on android-desktop
This "external/wpt/visual-viewport/scroll-event-order.html" test appears
to be the only one in the android filter that behaves differently on
android-desktop. Namely: it's expected to fail on normal android, but
passes on android-desktop. So this updates the expectations so that the
test is expected to pass or fail.

This should make the suite go green on the CI bot:
https://ci.chromium.org/ui/p/chromium/builders/ci/android-desktop-x64-rel-15-tests

After which we can promote out of experimental.

Include-Ci-Only-Tests: chromium.android.desktop:android-desktop-x64-rel-15-tests|android_chrome_wpt_tests
Bug: 436488951
Change-Id: I30a54669289884e341c2bdf7b3ca19c0adb00474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6847180
Reviewed-by: Jonathan Lee <jonathanjlee@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501042}
NOKEYCHECK=True
GitOrigin-RevId: 9f948faae37f3617d3d8aea37e708905711adbe1