[CDP] Add Target.TargetInfo.parentFrameId property
When auto-attaching to OOPIFs, it is essential to know where the
out-of-process frame sits in the frame hierarchy.

One can rely upon Page.frameDetached event when the frame
transitions from the local to remote. However, upon connecting
to a page with an existing OOPIF, the information about
the remote frame is never exposed through the Page domain.

The straightforward solution is to expose the parentFrameId
on the target, so it can be placed into the hierarchy.

Bug: none
Change-Id: I4264043343ed113d8798c2c5cf7c058d3ab2c2ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6842035
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501510}
NOKEYCHECK=True
GitOrigin-RevId: 7a31aa7eb1c6e8b4626b38aaf4a7108c9485e39e