[line-clamp] Don't clamp by lines and height with `CSSLineClamp` disabled
Recently https://crrev.com/c/6757119 added the ability to line-clamp
by both a number of lines and a height, whichever came earlier, if
`line-clamp` is set on an element where either `height` or
`max-height` are specified. However, that CL unintentionally also
enabled that behavior for `-webkit-line-clamp` even when the
`CSSLineClamp` runtime feature is disabled.

We have recieved a bug report for this. This might indicate that this
behavior with the `CSSLineClamp` feature is not web compatible, but
that is not necessarily the case, since the current behavior of
`-webkit-line-clamp` with the feature disabled is different from when
it is enabled in other ways (for example, lines after the clamp point
are shown when the feature is disabled, and hidden when enabled).
Therefore it makes sense to fix this unintentional exposure for now.

Since there are at least some sites that rely on the behavior of
`-webkit-line-clamp` not changing in the presence of `(max-)height`,
this CL also adds a use counter to track such cases, as well as use
counters for `line-clamp: auto` and for clamping by lines and height.

Bug: 40336192, 436345865
Change-Id: I5084073ed4aa3f5592110824019e70f37e6b7ffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818370
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501624}
NOKEYCHECK=True
GitOrigin-RevId: 0bf28d7654e16a1d1ab27e022fce0d3a29c5ac87