[Gap Decorations]: Disentangle Computed -Width Values from -Style
This CL decouples the computed values of border-*-width, outline-width,
and column-rule-width from their corresponding *-style properties.
Previously, when the style was set to none or hidden, the computed width
would resolve to 0px. However, following the CSSWG resolution[1], the
computed *-width values should now be independent of the *-style.

This CL also updates helper functions in
`CssSelectorFragmentAnchorTest`. Specifically, the `IsElementOutlined`
function was checking the computed value of the outline property. It is
now updated to check the resolved value instead.

Tests that previously expected none/hidden styles to yield 0px widths in
the StylePropertyMap have been rebaselined to reflect the updated
behavior.

Changes in this CL are behind an experimental feature flag.

[1]:
https://github.com/w3c/csswg-drafts/issues/11494#issuecomment-2675800489

Bug: 393631108
Change-Id: Ica29ac325e83b5e4f644ee225824965919f0c38c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6489213
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1501628}
NOKEYCHECK=True
GitOrigin-RevId: b3d149835549497c1c9f76b5e60c87493867af02