Run most WebCrypto tasks on the calling thread
With the exception of RSA key generation and sort-of[*] PBKDF2, none of
these operations have any reason to be moved to a background thread.
Cryptography is CPU-bound, so we need to the computation one way or
another. It is analogous to Array.sort(), not fetch().

It is also fast. We measure speed of symmetric cryptography in GB/s, and
we measure the speed of public key cryptography in thousands and tens of
thousands of qps. This means the thread hop is comparable, if not
higher, in cost to the operation itself.

This change does not fully mitigate WebCrypto's performance issues; the
promises infrastructure itself also imposes high costs that don't make
sense for these operations. (Again, the analogy is Array.sort(), not
fetch().) We also pick up a bunch of overhead from other API mistakes
(algorithm identifiers as complex structures, the spec-mandated copy of
all inputs). But this removes an overhead that wasn't forced by the API.

On a random benchmark on my machine, this switches us from taking 1.3
seconds to perform 100,000 SHA-256 operations in a loop, to taking 0.2
seconds. For reference, the underlying SHA-256 implementation would take
0.003 seconds, so we still have quite some overhead from WebCrypto
itself.

Right now this is implemented by just checking for two problem
algorithms. Better would probably be to wire the sync/async decision
into the algorithm implementation, but I've done the easy thing for now,
since we still have to have a base::Feature escape hatch. Once that
escape hatch is removed, we can trim this code a bit, and set up
something else.

Although they're now unnecessary, I've also kept the trace events from
before. DevTools tests seem to depend on this, so we're probably stuck
with it as public API now.

[*] PBKDF2 with large iteration counts is still fairly fast but starting
to exceed interactivity budgets depending on the iteration count. PBKDF2
with absurd iteration counts can take a very long time. We probably
should be rejecting those, but ah well.

Bug: 40842931
Change-Id: I7aa2910464c50bd42738da9e693534e7dcf5aee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6767493
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501639}
NOKEYCHECK=True
GitOrigin-RevId: 08e592f2700bbe6e6e782e56a571591c2ac315ed