Copy overlay color on LocalFrame swap
With RenderDocument, we'll change LocalFrames on every cross-document
navigation. Frame color overlays are saved in LocalFrame, and thus
needs to be carried over to the new LocalFrame to be preserved.
This CL makes it so that we do that by saving the previous
LocalFrame's  overlay color in OldDocumentInfoForCommit and applying
it to the new LocalFrame.

This is needed to make this test pass with RenderDocument enabled:
compositing/visibility/overlays-persist-on-navigation.html

Bug: 40615943
Change-Id: Iea0cb5c277f63004366487ac78f4a2598ebae931
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6728715
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501786}
NOKEYCHECK=True
GitOrigin-RevId: 33e7e40997835277faba6a2a22e2f5101ce1bb50