Menu: add long timeout to menuitem-activate.html
Per https://crbug.com/431038529, this test is timing out. I suspect it
is because there are many promise tests, each using the test driver to
perform a click, which I've observed being slow in the past. This CL
attempts to fix it by adding a meta long timeout to the test, and
removing the timeout from TestExpectations.

R=​jarhar

Bug: 431038529
Change-Id: I4819bf26bb94061ea3bff7153c3d96f7822bddcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6849348
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501933}
NOKEYCHECK=True
GitOrigin-RevId: bc01900600ba8cee10e3b94e94cd9791cab12e90