Use overlay scrollbar constants header directly in cc/.
Instead of redefining the constants, just use them from the header that
provides them. For clarity, this also renames the header to omit
"_aura".

The dependency is limited to just unittests because production code
should not be using these directly -- the sizes should be plumbed in
via the WebThemeEngine.

(There are pre-existing presubmit warnings for this header I will
address in a future patch.)

Bug: 431953872
Change-Id: I874496cdf59eaa46aa20afc58500a5e8f9579bb2
Bypass-Check-License: Renamed file
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6855754
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1502854}
NOKEYCHECK=True
GitOrigin-RevId: 50c6edc61932e0fee32397101f883fe0163805c8