Respect min height in BrowserControls::UnreportedSizeAdjustment.
If BrowserControlsParams::browser_controls_shrink_blink_size == false,
it means the blink::WebView is sized to the "large" viewport. But if we
also have top_controls_height == top_controls_min_height (and no bottom
controls), then the large viewport is the same as the small viewport.

The code that sets VisualViewport::browser_controls_adjustment_ didn't
understand this, so the VisualViewport wrongly believed itself to be
scrollable during a subframe-initiated ScrollRectToVisible operation.

This bug only existed on the main thread and not the compositor thread,
so the compositor would clamp the scroll offset back to 0 in the next
animation frame, creating a flicker effect.

Bug: b/437820869
Change-Id: Ie9f64a23f6e0cd9636f2f576fded69c8c98e5bbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6855060
Commit-Queue: Steve Kobes <skobes@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1503034}
NOKEYCHECK=True
GitOrigin-RevId: 317b0f59e5e2aa858090fb14c82a6b608c9eec22