Remove CustomizableSelect flags
This patch removes the following flags now that the origin trial and
enterprise policy to disable CustomizableSelect have expired:
- SelectParserRelaxation
- SelectParserRelaxationOptOut
- CustomizableSelect

There is an autofill test which was disabling the SelectParserRelaxation
flag. This patch rebaselines that test in order to make it pass without
relying on the SelectParserRelaxation flag.

Bug: 40146374
Change-Id: I39d378c1e8159254813606a831ee73cce56779f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6821896
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1503070}
NOKEYCHECK=True
GitOrigin-RevId: 2285459d5c394c07219ee90b49cf6dc2bdfac49d