Resolve the computed values in CSS Gradient functions
In order to properly resolve the computed value of the different values a CSS Gradient is composed of we need an instance of the CSSToLengthConversionData in order to compute the Lengths that depend on other style variables (eg, "em" units) or "calc" functions.

The best way to do this is in a new function of the StyleResolverState, an call it when the CSSToStyleMap is being computed to build the CSS Gradient value.

Bug: 40620723
Change-Id: I0b44b165d2b51bc801b5762d51314cefb6040e5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6676851
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1503281}
NOKEYCHECK=True
GitOrigin-RevId: f675a097d4768b6e83169be47b8f796725b287de