[Animated PB] Plumb load progress updates to BrowserControlsOffsetManager
We treat the progress bar as part of the browser controls.
BrowserControlsOffsetManager already has logic to move/animate the
controls, so we're also going to let it animate progress bar updates.

The animation is done via OffsetTags, and future CLs will follow for
managing the OffsetTag for the progress bar, and implmenenting the
animation logic.

Bug: 434769115
Change-Id: I6bb3534b694885d1eedc4be32e22f5f450d2b8d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825465
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1503479}
NOKEYCHECK=True
GitOrigin-RevId: cc8c1e7347befcb22b26e0a5e7a91eb71b902706