Convert non-compliant initiatorType values for resource timing
Some of the enum values of "fetch_initiator_type_names" are not
compliant value of `initiatorType` for resource timing. But they are
exposed directly in resource timing.

The new conversion converts them to "other" so the result is compliant
to the spec. Without the new conversion, the added tests would fail with
non-compliant value "use" and "attributionsrc".

Bug: 427115601
Change-Id: I8db7e8ee5c6acdf05cec5d255a08c3444dec5909
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6838381
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Jonathan Lee <jonathanjlee@google.com>
Commit-Queue: Guohui Deng <guohuideng@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1503586}
NOKEYCHECK=True
GitOrigin-RevId: fcf9e84dd7a683b7e4cc722bfdcd4b672e3d0014