[FontConfig] Pick fontconfig.h from correct directory
The `fontconfig/fontconfig.h` was moved from the upstream sources to a
local copy that is generated during rolling the library. In order for
that file to be found correctly, instead of being picked in the form of
an outdated version that exists in the sysroot, add the correct include
directory to the build file.

Fixed: chromium:439788709
Change-Id: I9d9c5f73f2e884650bdb43e1ec5266636dc74070
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6864179
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1503948}
NOKEYCHECK=True
GitOrigin-RevId: a1aa9f4a6047f6fe790caa545e75ba6dd23ab233