[Masonry] Update indexing to get intrinsic repeat track index
In the track collection, we were keeping track of the track index where
an intrinisic auto repeat started. The logic for calculcating this
didn't take into account that the indexing started over at each range of
sets. Update the logic to store this as a set index instead of a track
index, since it is more accurate, and assign it to the set index that
starts the whole auto repeat.

Also, scope this to auto repeats since we only care about grabbing an
index in this case.

Bug: 438215710, 343257585
Change-Id: Ibc8aaf3b760df9b632854511cce0640ba7a28677
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6847420
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1504184}
NOKEYCHECK=True
GitOrigin-RevId: 8862972e8d4c3a7bbef970059ccd9dfe0c6803e8