DevTools: Associate Media Inspector Players with their Owner DOM Node
This CL associates a MediaPlayer in the DevTools inspector side with its
owner DOM element (<video> or <audio>). This enables the frontend to
link to this DOM element from the player and vice versa.

The CL introduces a new AddDomNodeIdForPlayer() method to the
MediaInspectorContext interface, which is called by the MediaFactory
immediately after a player is created. The dom_node_id is now stored
within the MediaPlayer struct, and the DevTools protocol was updated to
include this ID when sending the list of players to the frontend.

******* CDP modifications ********
- Added Media.Player type
- Added Media.PlayerCreated event
- Removed Media.PlayersCreated event

Bug: 40269197
Change-Id: I03ed4c93d13af334e4c3f63fe4cc52346a941452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6853902
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1504300}
NOKEYCHECK=True
GitOrigin-RevId: cee463e9d7c44b1c2a934b5d04827f1446cc3602