Don't allow unsupported CSS math functions in CSSOM
Also, this CL enables parsing of anchor(), since we should discard it
only when we apply the expression to some property that doesn't support
it. However, for now it will be also disabled, since its CSSOM
representation is not implemented yet.

Fixed: 439560296
Change-Id: If3e7ddda5f27bdf8f5431f4ae4b12cdcae29ac8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6862763
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1504951}
NOKEYCHECK=True
GitOrigin-RevId: 04220e12bf3374a0913e2e499a79565b8d5abefd