[animation-trigger] Propagate triggers up the fragment tree.
The new proposal[1] for animation-trigger is that it declares a name (a
dashed ident) which is to be matched against the names declared by
trigger-instantiating properties, i.e.timeline-trigger, event-trigger
properties. The visibility of the names declared by
trigger-instantiating properties follows a model similar to anchor-name
- global by default, with the option of limiting the visibility with
a (yet to be implemented) scope property.

We are implementing this in 3 phases:

1. Compute AnimationTriggers from timeline-trigger declarations
(crrev.com/c/6813829 ).

2. Propagate the AnimationTrigger declarations from step1 up through the
fragment tree. (This patch)

3. Attach the AnimationTrigger objects to the associated Animations
after step 2 is done in layout (crrev.com/c/6775141 ).

This patch achieves the visibility of CSS trigger names by following the
Blink's anchor-name handling pattern - propagating the relevant
information up the fragment tree.

[1] https://github.com/w3c/csswg-drafts/issues/12336

Bug: 429392773, 390314945
Change-Id: I8e06a20f62aa5f7faa0ae887f348c3029c8d97a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6767352
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505163}
NOKEYCHECK=True
GitOrigin-RevId: 23d75b285c78eea39944613d52e378da4a4661b4