[Gap Decorations]: Introduce Grid Spanning Items Logic
This CL introduces the concept of "blocked cross tracks" to determine
which parts of a gap are obstructed by a spanning item. The
implementation begins with the creation of a `TrackRange` struct, which
denotes the start and end of a track spanned by an item. Each gap is
then associated with a `TrackRanges`, which represents a list of tracks
for which a gap is blocked.

In the absence of spanners, a column gap can be considered a contiguous
block starting from row 1 up to row N. The `blocked_cross_tracks` is a
set of all rows spanned by an item, thereby covering the gap. For
example, [(2,4)] indicates that the column gap has row 2 to row 4
blocked, resulting in the visible gap being [1,2) and (4, N].

Blocked tracks are collected during item placement using the
`AggregateGapsToBlockedTracks()` function. This function adds blocked
cross tracks for each gap that exists but is spanned or covered by an
item.

A unit test is added to exercise code changes introduced in this change.
Future changes will involve filtering gap intersections using these
built structures, which will be dependent on `*rule-break` property.

Bug: 436140061, 357648037
Change-Id: I7005ec9842019bbfc647ddcd4dcff083882139f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6870201
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1505352}
NOKEYCHECK=True
GitOrigin-RevId: 456910757899e13065c40c33ef4cca16dc1079dd