TaskAttribution: Don't propagate task state for sync events
We don't want to propagate task state if it would overwrite existing
task state, which is mostly a concern for dispatching synchronous
events. Overwriting state risks overwriting web scheduling state (spec
violation), and not propagating in that case is aligned with the current
direction of AsyncContext. Previously, we handled this on an individual
API basis, but that approach is brittle and isn't scalable.

This CL adds a new CreateTaskScopeIfTopLevel() method that sets up
propagation only if there isn't JavaScript already running, and it
converts a few places doing a bespoke version of this (and one that
should have).

The implementation uses v8::Isolate::InContext() to determine if JS is
running. This has the downside that entering a Context causes the method
to return true, so ordering between TaskScope and ScriptState::Scope
creation matters. If this becomes an issue (tests will catch this),
we can consider adding a v8 API for if there's a current stack frame.

Alternatives considered:
 1. Use IsRuningMicrotasks + TaskScope depth (including "null scopes"
    for callbacks, if needed). This gets us close, but there are ways of
    running JS outside of what we've instrumented, and instrumenting
    everything would be complex (e.g. some places call v8 functions
    directly)
 2. Propagate a special null context and only propagate if there is is
    no current task state. This is similar to (1), but assumes there's
    always a context being propagated. This has the same holes as (1).
 3. Handle this for each API. This would work, but it's challenging
    (some events can be sync or async, e.g. readystatechange, error,
    etc.), and is more brittle than handling this at the lower layer
    (Note: as a follow-up, I'm planning to remove the "unsafe" version.)

Bug: 40265789
Change-Id: Ib47bee03d7c5d63fd33119a0e6df50e0406b0137
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6831990
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505397}
NOKEYCHECK=True
GitOrigin-RevId: 2431ae8f48148a27223a32ca3572a63c3ebd7068