Caret-shape: Check whether painting caret by |PhysicalBoxFragment|.
In the existing overflow code, when caret shape is not bar, we call
ShouldPaintCaret() with |LayoutBlock|. This is not efficient as
every fragment in a block containing the cursor will have the
overflow added. This CL identifies the particular fragment for
the caret overflow.

Bug: 439706729
Change-Id: I9668372354cabd8a1570030bb810846a288a14cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6870571
Commit-Queue: Ziran Sun <zsun@igalia.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505733}
NOKEYCHECK=True
GitOrigin-RevId: fd6f1374e479299a7932c7906df3cdb96d5042a5