[SRI Message Signatures] Ship it.
Intent to Ship: https://groups.google.com/a/chromium.org/g/blink-dev/c/R6isgRnE1j4/m/MZRDE_dEAgAJ

This is a re-land of
https://chromium-review.googlesource.com/c/chromium/src/+/6861223, which
was reverted due to linker errors on Windows: it looks like the template
code for runtime enabled features is a bit confused if `stable` features
have `implied_by`. I'll file a bug for that and address it in a separate
CL, but for the moment, simply removing the `implied_by` declaration
seems reasonable.

Enabled-by-default-reason: Shipping the feature.
Bug: 375224898
Change-Id: I8e5254636c70343b4f760680ac384df44ca42b6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6870631
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505880}
NOKEYCHECK=True
GitOrigin-RevId: a42a9eebc5b89fc277625af7a48ce907ec614f1f