Ensure SyncToken generation in CanvasResourceProviderSharedImage
This CL ensures that a SyncToken is always generated after the
WillDrawInternal() call in GetBackingClientSharedImageForExternalWrite.
This is part of a prerequisite refactoring to ensure SyncToken
generation in CanvasResourceProvider is always localized and easy
to reason about, so that later we can safely apply scoped SyncToken
access.

Bug: 40286368
Change-Id: I5565cbc80b3c1d55c435deff610deb00a5aaa58e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6874649
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505888}
NOKEYCHECK=True
GitOrigin-RevId: fafff632e0ac03e95c2dc7d707c5c4abac88d71f