[html-in-canvas] Only notify of paint changes under the canvas
This patch changes the fireOnEveryPaint ResizeObserver option to only
fire when the canvas or its descendants need to repaint. This is a
requirement for efficiency, because we do not want to issue resize
observations for any change or rAF loop on the page. This is also
required for security, because we do not want to reveal paint
information for content outside the canvas.

Fixed: 440350908
Change-Id: I33342b06c73baf8e3ff689db4d0a236bf39af631
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6872465
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506110}
NOKEYCHECK=True
GitOrigin-RevId: f1b02ddbee24279aabdfb210b67b0c518ec75af9