Improve customizable select implicit anchor counting
There is a test crashing due to the select element's autofill popover
decrementing the implicit anchored element count below zero. This patch
replaces the logic to do implicit anchored element counting in the
InsertedInto/RemovedFrom methods of the autofill popover and the picker
popover with code in the constructor and pseudo-destructor of the
MenuListSelectType class, which fixes the crash and makes it count more
consistently.

Fixed: 439776357
Change-Id: I6f796426c39534d335815fe0013a633b1a682532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6862047
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506122}
NOKEYCHECK=True
GitOrigin-RevId: 320e43d06482fc22abe11ec391cb1dc2b7b8942e