Init FontCache with Default Skia Font Manager
Prior to this change, the FontCache always had a `nullptr`
`font_manager_`, except in 2 circumstances:

- On Windows, where the font manager is replaced with either the DWriteFontProxy or the FontDataManager. In both cases, the Skia Default Font Manager is also replaced at the same time, so `font_manager_` was always equal to the Default font manager.
- In some tests (when `features::kFontDataServiceAllWebContents` is set), in which case the default skia font manager is also overridden.

The FontCache and other related code had to check if `font_manager_` was
set before using it, and fallback to the skia Default Font Manager if it
wasn't.

This change initializes the `font_manager_` field to the skia default
font manager, which is equivalent to the previous behavior but
simplifies the FontCache code.

crbug.com/433293142 and similar bugs have been popping up for many
years, and make it difficult to assess the rollout of the new Windows
Font Data Service. Simplifying this code will make it easier to debug
this class of issues.

Bug: 433293142,40260875
Change-Id: I47d39c30894529fbf254ef7d2a2f21b727e0625b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6871519
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Kaylee Lubick <kjlubick@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506441}
NOKEYCHECK=True
GitOrigin-RevId: 59fb11ca726714a6aef62eb01b67d4fd491e47bf