[CSP] Clarify report-only console messages.
Based on feedback from developers, including [1], the current format
of report-only CSP violations is more confusing than it could be. This
CL rewords the console message in a way that's hopefully clearer about
both the violation and its impact on the page.

[1]: https://github.com/w3c/webappsec/blob/main/meetings/2025/2025-08-20-minutes.md#:~:text=%22%5BReport%20Only%5D%20Refused%20to%20load%22%20%3D%3E%20still%20loads%3F!

Bug: 441128473
Change-Id: I0a31dc0684ddc2b93edaa13c4afe715b90ae715f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6870550
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506471}
NOKEYCHECK=True
GitOrigin-RevId: 9943d05853907c0933d72130228211d50f4dca92