Assert that RTCVideoEncoder::InitEncode doesn't change codec
This makes it explicit that we assume RTCVideoEncoder::profile_ can be
initialized upon construction and never changed. Codec type changes in
the real world are expected to involve recreating RTCVideoEncoder.

We can now remove a check in CodecSettingsUsableForFrameSizeChange()
that can never be true, and a test that exercised an unrealistic
scenario.

Bug: 435583653
Change-Id: I69c90dba1a6c9268daea6ef20917d186303dd6be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6858775
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506739}
NOKEYCHECK=True
GitOrigin-RevId: 684a82c0e0996dd7c1538fabd4c6391e237b10ae