[anchor] Keep anchored elements within their fragmentation context.
Don't use `StitchedAnchorQueries`, or anything at all in
anchor_query_map.cc

If the containing block of an out-of-flow positioned element
participates in a fragmentation context, so does the out-of-flow
positioned element. This was violated for anchor-positioned elements. In
such cases we'd essentially let them "escape" the fragmentation context,
take a peek on the outside, to calculate the bounding box of the anchor,
and then jump back into the fragmentation context. This caused undesired
behavior, made it hard to reason about, and also caused implementation
complexity.

Anchor-positioned elements use the `anchor()` function with inset
properties, such as `top`, `right`, `bottom`, and `left`. These are
relative to the containing block for out-of-flow positioned elements,
and anchor-positioned elements should be no different. Therefore, if the
containing block is fragmented, we're in this "stitched" coordinate
system, where we imagine that all fragments have been stacked on top of
each other. This is also the case for sizing properties, such as
`block-size`. `<div style="columns:2;"><div style="block-size:100px;">`,
for instance, will not create a 100px tall child, visually, since it
will be fragmented into two columns, with 50px of content in each. Yet,
it will be 100px tall if we imagine the fragments to be "stitched"
together. If we now imagine that the `block-size:100px` DIV is
relatively positioned, and put a `position:absolute` child inside with
`top:70px`, it will not start 70px into the first column (its containing
block is fragmented). Nay, it will start 20px into the second column,
skipping the 50px of the first column, then skip the remaining 20px in
the second column. This hopefully explains the weirdness with the
previous behavior.

The fix in PhysicalBoxFragment::MutableForOofFragmentation::Merge()
should have been there in the first place, since we need to handle
anchors with the same name correctly. Previously, we'd be saved by the
duplicate logic in StitchedAnchorQuery::AddAnchorReference(), which
would clean up after this mistake. Now we need to get it right the first
time. anchor-position-multicol-012.html (and some existing, rather
convoluted, tests) would fail without this.

Update tests to match the new behavior. Changed them to reftests, since
offsetLeft & co (used by check-layout-th.js) have poor interop inside
multicol. See https://github.com/w3c/csswg-drafts/issues/11541

This change is behind a runtime enabled feature kill-switch named
CSSAnchorSimplifiedFragmentation.

See https://github.com/w3c/csswg-drafts/issues/12287 for the spec
discussion, and resolution.

Bug: 436305267
Change-Id: I4872d35a040062e7dbccaf736b9eda3b8661b880
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6590416
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506964}
NOKEYCHECK=True
GitOrigin-RevId: 830f9cf847e3e353260c44e1d9c929a59ba84af4