[SRI Message Signatures] Reduce test flake.
The inspector tests for SRI Message Signatures wait for a single issue
to be added, but this ends up racing in a way we periodically lose
against an issue regarding a blocked resource.

This CL narrows the tests to wait for and return a
`SRIMessageSignatureIssue` rather than any issue.

Bug: 381044049
Change-Id: If68e4539ae5c6e3bd57a79e98e2ed756652b4391
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6884983
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506966}
NOKEYCHECK=True
GitOrigin-RevId: f64ae041ae67b6bb165fdbab0d80a7f7286be2da