EventTiming: Fix Failing CHECK and ensure timestamp before using it.
We recent changed to have every event dispatch, regardless of type, set
the enqueue time before EventTiming reads this value as part of
`WidgetEventHandler::HandleInputEvent`.

We expected there was a chance that this would not be 100% coverage
because there might be code pathways that EventTiming observes, but
which aren't going through the same WebInputEvent dispatcher where we
read and save the enqueue time. One hypothesis right now is IME input.

This patch removes the CHECK and re-adds a guard on the value before
using the timing. A followup task will be to either find and resolve the
issue, or to add histogram/usecounter that help find the situations
where this happens.

Bug: 422215352,440723922
Change-Id: I169f6ca364b4052642a613b65beaa550f40f92db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6892274
Commit-Queue: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507335}
NOKEYCHECK=True
GitOrigin-RevId: 5753c0d66ef34314d897f6db16d61d8a53ce563d