[SVG] Fix error handling for SMIL animation value parsing
According to SMIL[1], if any of the values provided via the from,
to, by or values attributes are invalid, the animation should have
no effect. Similarly, if none of these attributes are specified, the
animation should have no effect.

Currently, we do not properly catch parsing errors, and the animation
continues for the portion that is valid. This CL improves error
handling by ensuring that parsing failures are detected and the
animation is suppressed when required.

[1]: https://www.w3.org/TR/2001/REC-smil-animation-20010904/#AnimFuncValues

Bug: 40731437
Change-Id: I931c27f26e200d41077f0a6c6e382b217a9c4336
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6491666
Reviewed-by: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Virali Purbey <viralipurbey@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1507663}
NOKEYCHECK=True
GitOrigin-RevId: a270f982f7fc680315c91d0ffa8130c07a5f3ef4