Flip the direction of surrogate_for for line-break
Surrogates with different sets of valid values do not behave correctly.
For line-break / -webkit-line-break, one is not a subset of the other
for valid values. That means we may return computed values which do not
round-trip. Additionally, only the main property of surrogates is used
for var() substitutions.

For the line-break property, variable substitutions would be done as if
we were using the -webkit-line-break property, which meant the
standardize 'anywhere' keyword would be dropped.

Flip the surrogate_for around to make line-break the official property.
That will result in var() substitutions fail for 'after-white-space',
which is more acceptable.

Bug: 441536815
Change-Id: I019fe0f84a8a42287b6aa61368fe06d60254a433
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6895238
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507669}
NOKEYCHECK=True
GitOrigin-RevId: 28caa3e87912f2edc9e07e4d910f9e3618f42c19