Enable SystemLoopbackAsAecReference by default
This required changing a bunch of tests which assumed that loopback AEC
was disabled. The difference stems from the fact that loopback AEC takes
precedence over platform AEC when requesting EchoCancellationMode::kAll.

It also means that some test cases related to
LocalMediaStreamAudioSource had to be bypassed entirely, because they
were testing a scenario when LocalMediaStreamAudioSource would never be
created to begin with. This is done in the following files:
* local_media_stream_audio_source_test.cc
* media_stream_constraints_util_audio_test.cc

the GetUserMediaEchoCancellationModes runtime flag, and will be part
of that launch.

Enabled-by-default-reason: The functionality is still guarded behind
Change-Id: Ib00a2b92bd1e9ce44895870749d7739f2a537212
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6884621
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Fredrik Hernqvist <fhernqvist@google.com>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507712}
NOKEYCHECK=True
GitOrigin-RevId: 0137d6b50b9b98d96feb8fa90a97fca3ad20a6b8