Make DOMMatrix invertSelf() tests handle -0 as identical to +0
Because the spec[1] treats -0 and +0 as equal, this patch uses
checkMatrix() instead of checkDOMMatrix(). checkDOMMatrix() compares
them strictly, while checkMatrix() uses a small epsilon and treats them
as equal.

The test failed because a 4x4 identity matrix inversion can end up with
-0 entries. Unlike Firefox and WebKit, Chromium uses a 2D fast path in
gfx::Matrix44::GetInverse() if the 4x4 matrix can be represented as a 2D
transform. As a result unary operations in that path can turn 0 into -0.
The 2D path is an intended optimization. So, the logic is correct.

This keeps the 2D fast path. It just fixes the failing test without
changing runtime behavior.

[1] https://www.w3.org/TR/geometry-1

Bug: 357931029
Change-Id: Ia22abf44d14289693832cb2d7493599499d12244
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6880947
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Jinho Bang <zino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507860}
NOKEYCHECK=True
GitOrigin-RevId: 3b113074994a49df1b4af475e559a2eb9411355c