[line-clamp] Refactor `InlineLayoutAlgorithm::GetLineClampState`
The `GetLineClampState` method of `InlineLayoutAlgorithm` returns
whether the current line is affected by `line-clamp` or
`text-overflow`, and how. This is used to determine whether to place a
`LineTruncator` ellipsis, or whether to hide the line.

This method takes a pointer to a `LineInfo`, which is used to
determine whether the `text-overflow` ellipsis needs to be placed
depending on whether the line overflows, or whether a `line-clamp`
ellipsis needs to be skipped because the line is a block-in-inline.

However, if this pointer was not mandatory, and the post-line-breaking
information was ignored, then this method could also be used to
determine whether to place the `line-clamp` ellipsis as part of line
breaking, which will become more relevant when the `block-ellipsis`
property is added, which will allow skipping the `line-clamp`
ellipsis.

This method also takes an argument for the line box height, which is
no longer used. This patch removes that as well.

This is a refactor that should not change behavior, and therefore it
adds no tests.

Bug: 40336192
Change-Id: I73c62ed8cfb5905fb9da5a74415dc8232c1effad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6888530
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1507881}
NOKEYCHECK=True
GitOrigin-RevId: 4faac46c143f23f6574e697ed99722de52b4e852