[line-clamp] Implement the `block-ellipsis` longhand
Previously, when making `line-clamp` a shorthand property
(https://crrev.com/c/6760402), we purposefully didn't implement the
`block-ellipsis` longhand. This is because there was an ambiguity in
the shorthand's syntax, and the way to resolve that seemed to be to
remove the `none` keyword for `block-ellipsis`. However, since we were
not planning to ship `block-ellipsis: <string>` at first, this would
leave that property being useless in our implementation, since it
would only have one possible value.

However, the CSSWG has resolved to instead rename the `none` keyword
to `block-ellipsis: no-ellipsis` (bug 440063559). This removes the
blocker for implementing this property in Chromium (since now it has
two possible values: `auto` and `no-ellipsis`), and so this CL
implements it.

This patch also fixes and adds some related WPT tests.

Bug: 40336192, 440063559, 440951911
Change-Id: I614c48a5d27bb23e4beb7c8fb0b51ea660be1b11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6888571
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1507886}
NOKEYCHECK=True
GitOrigin-RevId: a4e05b8b1e5a1d5d8417e4b5bb1f28edd64dcbaf