Canvas 2D globalHDRHeadroom: Enable PQ and HLG base gainmap image tests
These were disabled because of incorrect rendering of HLG and PQ ICC
profiles.

Allow some fairly significant error for the PQ version (4% in sRGB
space) because both the PQ image and the gainmap image have substantial
quantization error (it may be that I can re-make the jpeg images to at
least have the error in the base image be compensated for in the gain
image).

Bug: 428575083
Change-Id: I150b19dff74b1fbd04eb027412ad2f83f003a3d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6895248
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507960}
NOKEYCHECK=True
GitOrigin-RevId: 615acc5230f1f26c26c1e5a13071fadf1b373a9a