Change the setPermission protocol commands to also accept an embedding origin
This is a step toward correcting the flawed assumption that all
permissions can be keyed on a single origin. This change adds
'embeddingOrigin' to the protocol commands, and now considers 'origin'
as the requesting origin. Next, the WebDriver permissions plumbing will
be updated to also pass an embedding origin.

Design doc:
http://docs/document/d/1n5OcaG41h4IDvNmlTA18kCLQhyLmpS-DU_q615JvDd4?pli=1&tab=t.0

Bug: 384959114,434724810
Change-Id: I3ef2fd8b58c530eb4863a5f1415fb50d975f8f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825262
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Elias Klim <elklm@chromium.org>
Commit-Queue: Jalon Thomas <jalonthomas@google.com>
Cr-Commit-Position: refs/heads/main@{#1508174}
NOKEYCHECK=True
GitOrigin-RevId: 749c713ff33f4b7759f140bab6c92519d1c08984