Ensure SyncToken generation in OverwriteImage
This CL ensures that a SyncToken is always generated after the
WillDrawInternal() call in OverwriteImage().

This is part of a prerequisite refactoring to ensure SyncToken
generation in CanvasResourceProvider is always localized and easy
to reason about, so that later we can safely apply scoped SyncToken
access.

Bug: 40286368
Change-Id: I4999e26720cabffd7daa54efba1c6e0426506654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6891434
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508367}
NOKEYCHECK=True
GitOrigin-RevId: dcc531a948f5e2d03a8b5f079c34e96cac0cbf25