Debug detached WindowProxy script execution
Per https://crbug.com/435149285 the AndroidReaderModeUseReadability
experiment appears to be causing a CHECK failure [1] in the following
flow:

logging::CheckFailure()
LocalFrame::RequestExecuteScript()
LocalFrameMojoHandler::JavaScriptExecuteRequestInIsolatedWorld()

This is caused by `ToScriptState()` unexpectedly returning nullptr,
which should never be the case when invoked by this mojo method, as the
validity of the mojo invocation should be correctly bound to the
lifetime and validity of the execution context and its ability to
execute script (i.e., being attached, among other things).

This CL captures a lot of scoped crash key information about the frame
in this case, and dumps without crashing instead of actually crashing,
since that information isn't useful.

An earlier version of this CL plumbed a bunch of callsite-specific
information through the mojo call and into the renderer, but this was
discarded as unnecessary when we got clarity on the exact callsite that
was definitely causing the crash. Per [2] and [3] the DOM distiller
component is injecting the "readability triggering" script into an RFHI
whose WindowProxy is apparently in some kind of non-initialized state
that causes ToScriptState() to return null in the mojo handler.

[1]:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/frame/local_frame.cc;l=3206;drc=37e21de1acb8c8061dd13dc41aaa22ea7a1ea992
[2]: https://crbug.com/435149285#comment10
[3]:
https://crash.corp.google.com/browse?q=expanded_custom_data.ChromeCrashProto.magic_signature_1.name%3D%27blink%3A%3ALocalFrame%3A%3ARequestExecuteScript%27+AND+product_name%3D%27Chrome_Android%27+AND+expanded_custom_data.ChromeCrashProto.channel%3D%27beta%27&compProp=expanded_custom_data.ChromeCrashProto.experiments.ids&v2=4b1af6a5-81f8e17d&v1=4b1af6a5-b09b3d6d

R=kevers@chromium.org

Bug: 435149285
Change-Id: I4a148f849753dcea42e2077b50b6b0644a011b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6861835
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508381}
NOKEYCHECK=True
GitOrigin-RevId: 5528f2e8194fe42e6f41724f2aa0f51d7ff6d34d