Remove libxml2 legacy compatibility code
We do not want to keep legacy compatibility code in our code base. When
downstream projects compile against external libraries, we can't
guarantee stability if the versions built and linked against are not up
to date.

This holds similarly for building against external not up-to-date
HarfBuzz for example, and we do not guard against that in font code.

Bug: chromium:441911594
Change-Id: I9e3c28373c547358ec07da2f05dfba46f521cae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6898007
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508510}
NOKEYCHECK=True
GitOrigin-RevId: 7a9c043250d5642870facdff79149dde7afff1d6